Updated a lot, see details
1. redid java docs to be a bit more descriptive.
2. Fixed interfaces to include details for fields and constructors.
3. added interfaces for server and client, but if I'm going to honest they are 100% useless. Might be important that we have them anyway
4. Updated test cases for interfaces to include server and client

Note: its possible some of the contents are a little different than others' local versions. I'm not sure how they could differ, but I've done testing to make sure all of these versions work, so I'm uploading them as a sort of way to make sure everyone is on the same page with contents.